# Conflicts:
#	nts.uk/uk.com/exio.ctx/nts.uk.ctx.exio.dom/src/main/java/nts/uk/ctx/exio/dom/input/canonicalize/domains/generic/IndependentCanonicalization.java

+++ nts.uk/uk.com/exio.ctx/nts.uk.ctx.exio.dom/src/main/java/nts/uk/ctx/exio/dom/input/canonicalize/domaindata/KeyValues.java
@@ -15,8 +15,8 @@ public class KeyValues {
 		return values.get(index);
 	}
 	
-	public int getInt(int index) {
-		return (int) values.get(index);
+	public Long getInt(int index) {
+		return (Long) values.get(index);
 	}
 	
 	public BigDecimal getBigDecimal(int index) {
